backend/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
google-generativeai
numpy
